Split Configure schemas into a separate file
This is to reduce the length of the Configure file. Also the error
file path has now been defined in the Constants.
